Update readme.md - variables block on the top
I found that if you use this to deploy several labs, you need to more quickly derive the variables used for this lab.

I have moved all required variables to the top that require adjustment and from here they can be used on the management host. Hope you appreciate this change.

In addition, this PR removes duplicates of $servers variable used during the lab.